[#121] Improve sourceMap performance
Writes the files directly to a source map generator removing a bunch of
intermediate steps performed when converting between nodes to a
generator.

On my laptop this brings the concat step of my build down from ~6m to
under 10s.

The current implementation does not seem to add mappings before filler
lines (banner, separator and footer) in a consistent manner so this
never adds a mapping for the filler text.